// editor/src/utils/api.ts
export interface FileDto {
  /** @format guid */
  id: string;
  path?: string;
  isImageFromSocial: boolean;
  fullPath?: string;
  originalFileName?: string;
  fullPathWebP?: string;
  fullPathJp2?: string;
  dominantColor?: string;
  type: FileType;

  /** @format int32 */
  width: number;

  /** @format int32 */
  height: number;

  /** @format double */
  size: number;
}

export enum FileType {
  Favicon = 'Favicon',
  SocialImage = 'SocialImage',
  Preview = 'Preview',
  TemplatePreview = 'TemplatePreview',
  TemplateFull = 'TemplateFull',
  ProfileAvatar = 'ProfileAvatar',
  ProfileCover = 'ProfileCover',
  Elements = 'Elements',
  Community = 'Community'
}
// ...
export interface BaseProjectPageDto {
  /**
   * Page identifier. For save changes in subsequent queries
   * @format int32
   */
  id: number;

  /**
   * Project identifier. Make for future if user will have several projects with one template
   * @format int32
   */
  projectId: number;

  /** Page name */
  name?: string;

  /** Mark page like homePage */
  isHomePage: boolean;
  url?: string;

  /** Open graph images */
  socialImage?: FileDto;
  previewImage?: FileDto;
  title?: string;
  description?: string;
  keywords?: string;

  /** Is not index page by Google, Yandex robots */
  isNoIndex: boolean;

  /**
   * Last edit date
   * @format date-time
   */
  lastUpdateUtcDateTime?: string;
  isPublished: boolean;

  /** @format date-time */
  lastPublishedUtcDateTime?: string;
  isHasUnPublishedChanges: boolean;
}

export type ProjectPageDto = BaseProjectPageDto & { settings?: string };

export interface CreateProjectPageDto {
  /** @format int32 */
  templateId: number;
}
// ...
export interface BaseProjectDto {
  /** @format int32 */
  id: number;
  name?: string;

  /** @format int32 */
  templateId: number;
  status?: ProjectStatus;
  statusName?: string;
  projectUrl?: string;
  settings: ProjectSettingsDto;

  /** @format date-time */
  lastUpdateUtcDateTime?: string;
  previewImage?: FileDto;
  isPublishedInCommunity: boolean;
}

export enum ProjectStatus {
  Draft = 'Draft',
  Published = 'Published'
}

export type ProjectSettingsDto = BaseProjectSettingsDto & { faviconImage?: FileDto };

export type BaseProjectSettingsDto = DomainSettings & {
  title?: string;
  description?: string;
  shortUrlLink?: string;
  googleAnalyticsId?: string;
  googleTagManagerId?: string;
  hotjarId?: string;
  facebookPixelId?: string;
  isAirtapBranding: boolean;
  brandingTheme: BrandingThemeType;
  brandingPosition: BrandingPositionType;
  projectUrl?: string;
};
// ...
export type PagedProjectDto = BaseProjectDto & {
  likes: number;
  views: number;
  hasLike: boolean;
  homePage?: BaseProjectPageDto;
  hasAnotherPublishedPage: boolean;
  pagesCount: number;
  categoryId?: number;
  tags?: number[];
};

export enum ProjectFilterType {
  All = 'All',
  Community = 'Community',
  Custom = 'Custom',
  Drafts = 'Drafts'
}

export interface CreatedProjectDto {
  /** @format int32 */
  projectId: number;

  /** @format int32 */
  pageId: number;
}

export interface CreateProjectDto {
  /** @format int32 */
  templateId?: number;
  isWithBlankPage: boolean;
}